requirements.txt
yfinance
matplotlib
python-telegram-bot
schedule
google-generativeai
python-dotenv